+++ app/src/main/java/woowacourse/movie/domain/movie/Movie.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.movie
 
 import java.time.LocalDate
