+++ app/src/main/java/woowacourse/movie/domain/theater/Seat.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.theater
 
 data class Seat(val row: Int, val col: Int) {
     fun price() = SeatGrade.of(this).price
