+++ app/src/main/java/woowacourse/movie/presenter/movieSelection/MovieSelectionContract.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.presenter.movieSelection
 
-import woowacourse.movie.view.model.MovieListItem
-import woowacourse.movie.view.model.MovieListItem.MovieUiModel
+import woowacourse.movie.view.model.movie.MovieListItem
+import woowacourse.movie.view.model.movie.MovieListItem.MovieUiModel
 
 interface MovieSelectionContract {
     interface View {
