+++ app/src/main/java/woowacourse/movie/presenter/seatSelection/SeatSelectionPresenter.kt
@@ -1,11 +1,11 @@
 package woowacourse.movie.presenter.seatSelection
 
 import woowacourse.movie.R
-import woowacourse.movie.domain.Seat
-import woowacourse.movie.domain.Theater
-import woowacourse.movie.domain.Ticket
-import woowacourse.movie.view.model.TheaterUiModel
-import woowacourse.movie.view.model.TicketUiModel
+import woowacourse.movie.domain.movie.Ticket
+import woowacourse.movie.domain.theater.Seat
+import woowacourse.movie.domain.theater.Theater
+import woowacourse.movie.view.model.movie.TicketUiModel
+import woowacourse.movie.view.model.theater.TheaterUiModel
 import woowacourse.movie.view.model.toDomain
 import woowacourse.movie.view.model.toUiModel
 import woowacourse.movie.view.seatSelection.SeatSelectionActivity
