+++ app/src/main/java/woowacourse/movie/view/model/movie/TicketUiModel.kt
@@ -1,8 +1,8 @@
-package woowacourse.movie.view.model
+package woowacourse.movie.view.model.movie
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
-import woowacourse.movie.view.model.MovieListItem.MovieUiModel
+import woowacourse.movie.view.model.movie.MovieListItem.MovieUiModel
 import java.time.LocalDateTime
 
 @Parcelize
