+++ app/src/test/java/woowacourse/movie/domain/DayTypeTest.kt
@@ -3,6 +3,7 @@ package woowacourse.movie.domain
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import woowacourse.movie.domain.schedule.DayType
 import java.time.LocalDate
 
 class DayTypeTest {
