+++ app/src/test/java/woowacourse/movie/domain/MovieTest.kt
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import woowacourse.movie.domain.movie.Movie
 import java.time.LocalDate
 
 class MovieTest {
