+++ app/src/test/java/woowacourse/movie/domain/SchedulerTest.kt
@@ -2,6 +2,7 @@ package woowacourse.movie.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import woowacourse.movie.domain.schedule.Scheduler
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.LocalTime
